requirements.txt
fastapi
uvicorn
python-telegram-bot
python-dotenv
requests
supabase
feedparser